pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.1"
ruff = "^0.5.4"
mypy = "^1.11.0"
pytest-env = "^1.1.3"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
